More changes to support >1-width runes.
The framed and edit widgets have been changed to better support wide runes,
and also to fix some utf8 problems. Some of the approaches used in edit
are poor, iterating over the length of the input (O(n)). I ought to refactor
and maintain (a) a before and after cursor structure representing the text
each side of point, and (b) store the data as []rune rather than string to
avoid having to decode the utf8 each time to get the runes.