Interpolate colors for RGB in 16-, 8- and 2-color modes.
As suggested by @R-Smith - this makes sure that a gowid program using
RGBColors in 16-, 8- and 2-color mode will not crash. Instead, go-colorful is
used to try to estimate the closing matching color from the limited palettes
that these modes provide. An alternative and more manual approach is to used
modeswap.Color. 16-color mode is untested, so probanly not working yet - I'll
test when I have a terminal that reports support for 16 colors.