Fixes the "double output" of utf-8 encoded runes in zsh.
This happens when gowid's terminal widget processes a title-changing control
sequence like '\e]2echo 中' - the non-ascii character was not included in the
title change, and instead emitted to the terminal as a regular
character. Because of the way zsh updates the title after each command, the
non-ascii character appeared twice on the terminal - once from the incorrectly
processed title change, and then again from the echo command.