Pile widget's last child can incorrectly claim user input.
This came from issue #17 via @R-Smith. Ryan's example shows the right hand
column taking focus when the mouse hovers over it. The problem stems from a
bug in the pile widget, also used in Ryan's example - the pile incorrectly
concludes that its last child widget takes the user input when the event is a
simple mouse movement, and this bubbles up causing the right hand column in
Ryan's example to be given focus too.