Fix propagation of UserInput() results through containers.
Thanks to @R-Smith for the bug reports. There was some incorrect logic in the
container widgets resulting in the behavior reported in #16 and #17. The
containers were changing the child in focus based on the respective child
accepting the mouse user input. But it's better to change the focused child
if (a) the child that is the target of the mouse event is selectable (designed
to take focus) and (b) the event is a left mouse click - what the user would
expect to do to affirmatively move focus with the mouse. The fib.go example
shows this - the children are Selectable() but don't process user
input. But the user would expect to be able to click on one to change the list
focus, even if the child widget doesn't do anything with the input.