New APIs to support starting and stopping a UI dynamically.
ActivateScreen() will have gowid/tcell take over the
terminal. DeactivateScreen() will undo that, returning the terminal to the way
it looked before the gowid application started. This is intended for
situations where you want to let the user go back to look at the terminal as
it was prior to the UI launching; OR, as in termshark, supporting SIGTSTP in a
sensible way (cf running emacs -nw in the terminal then hitting ctrl-z).