Bug fix - if a new model is provided for a table, clear the cache of widgets.
I'm not sure if this is always right yet... The idea is that changing
the table model invalidates all widgets cached based on the old
model. Widgets are cached because usually you would want to preserve the
state of a widget as you navigate around a table e.g. a specific row has
a checkbox widget that is checked. But a better model might be pushing
through widget state to the model somehow and skipping caching
altogether..?