If event sequence contains a run of mouse movement only, discard some.
This fixes #25. If the sequence of events features a contiguous run of
mouse movement events, with no button clicks, then I assume I can drop
the intermediate events. This is not ideal - for example, something like
a terminal painting app might want to receive every mouse event. A
better solution might be to drop events only if a backlog is
detected. For future consideration...