Provide a standard way to map from a table RowId to its displayed row.
Currently the simple table model provides a map from displayed row to
RowId. This allows the rendering routine to display the correct row if
e.g. the table is sorted - the table asks what to display at row #0,
then gets back the identifier for that row, id=567 (which might be the
row sorted into position #0). This inverted map provides a way for
clients of the table to focus on a displayed row given the the row's
identifier e.g. move table focus to displayed row #0 because I want to
focus on row id=567.