A better UX when clicking inside pile and column widgets.
Prior to this change, a down-click of the left mouse button inside a
pile or a columns widget would set focus to the clicked subwidget, if
that widget was "selectable". This is in contrast the button widget,
which waits for the click-release, and checks to see whether the widget
over which the mouse was clicked is the same one. This is more forgiving
and lets the user abandon a click by moving the mouse away before
releasing. This change makes the pile and column widgets line up with
that UX.