Fixes some scrolling issues in lists with multi-line widgets.
These showed up while implementing stream reassembly in termshark. To
see the effect in action, run the gowid table example in a terminal of
about 60 columns (so that some rows require more than one line). Use
cursor down in the first column until you reach the bottom of the first
page and beyond. Then slowly cursor up. Prior to this change, the rows
would occasionally shift up and down with each cursor movement. The
problem was an incorrect calculation of the ration of the top of the
current widget to the number of screen lines, in particular when the
multi-line child widgets have a different height when rendered with
focus.