Add a more useful callback registration method.
This method registers a callback that can accept parameters, provided as
a variadic array of interface{} - meaning the registrant understands and
can make use of the type of arguments sent by the issuer of the
callback. I plan to make use of this in the vscroll widget by adding a
right click callback which passes, as a parameter, how far down the
scrollbar was right clicked.