Extend the scrollbar widget with support for right clicks.
The idea is to allow for quicker access to scroll regions by right
clicking on different parts of the scrollbar - the further down, the
further the UI should advance through the data. Currently, left clicks
above and below the bar will issue a pgdown and pgup respectively -
which can make navigating through long lists of data too slow.