Prefer tcell's dynamic terminfo DB
This affects gowid's terminal widget only. I am making this change
because it gives the widget the best chance of using an up-to-date
terminfo database, at the cost of invoking infocmp. I expect this will
only be needed once, even if multiple terminals are spawned by the TUI,
since they'll probably have the same TERM variable, and the widget will
cached the resulting Terminfo struct.