Validate urwid colors at construction rather than when used
This change is motivated by a desire in termshark to fail a parse of a
color if it doesn't match one that is known. I then want to fall back to
a default color. But the parse won't fail because the urwid color
constructor will accept any string as an argument.