Be sure to set the screen's color range when the screen is activated
An optional argument to the gowid app prevents immediate screen
"activation" - that prevents the application taking over the screen,
clearing it, etc. I did this for termshark - so that I could write a
message to the original terminal to let the user that termshark would
launch its UI when packets were detected. But this led to a bug where
the screen's idea of how many colors it supported was not initialized.