Support parsing tcell's own color structs by name
MakeTCellColor() is intended to understand its argument as a string
representation of a color from the tcell package. Prior to this change,
this covered specially declared names in the tcell.ColorNames array. Now
with this change you can refer to the e.g. 123rd color in the 256 color
space as "Color7B". I am going to make use of this in termshark's
upcoming themes feature.