Make sure terminal color codes emitted aren't affected by COLORTERM
If this variable is set to truecolor, then gowid will assume (via tcell)
that the terminal supports 24-bit color, and when the programmatic
colors are translated to codes for the terminal, you will instead see
something like ESC[38;2;192;192;192;48;2;0;0;0m. This defeats the
purpose of the program which is supposed to show correct operation on
terminals via 16-color and 256-color codes.