Provide a way to adjust the rendering height of the menu widget
The menu is really just an overlay, and so this is a passthrough to the
overlay configuration. It's useful to change the height from the default
because if the outer rendering context is an IRenderBox, a fixed layer
will render without regard to the box's height constraint. This results
in the bottom of the menu being cut off when it's trimmed to fit in the
box. If you know the number of items in your menu is 16, you can then
call

m.SetHeight(gowid.RenderWithUnits{U: 16})

(for example)