The padding widget did not respect the vertical alignment offset
This widget is an amalgam of the vpadding and hpadding widgets. If the
rendering context is an IRenderBox, the subwidget should should rendered
with a box that is smaller by the amount of the vertical alignment
offset. Without this, padded widgets will be rendered with an overly
large height value that will result in their bottom being cut off when
displayed