Adjust the way subwidget render sizes are computed
This change makes sure that the contained widget is not rendered with a
size that is larger than the containing widget; it does this by choosing
the miniumums of columns and rows, when they are provided. An exception
is if the contained widget is rendered as "fixed"; then it may render
larger than its parent (but it will be truncated when the canvases are
assembled into the final screen).