Fix a potential nil dereference in the list widget
When a list widget subitem is clicked, the action is only registered on
release; the list widget has to navigate the underlying model (list
walker) to find the widget that should be activated on mouse button
up. But if the navigation results in end-of-list, a nil is returned from
the walker, and I didn't check the boolean indicating whether or not a
next (or previous) item was found.