Add selected focus support to the list widget
Both the column and pile widgets already have this feature. When a
composite widget is rendered, it will either be in focus or
not. However, it's useful in some UIs to have a composite widget
indicate the widget that would be in focus if the composite widget
itself had focus. Termshark uses this to show a less colorful indicator
of the currently selected packet, for example. To make this work, the
composite widget will call Render() on its non-focus children with the
selected flag set instead. The children can use something like the
isselected.Widget to style themselves appropriately. This change causes
the list Widget (and hence also the tree Widget) to set this selected
flag when rendering their non-focus children, too.