Rendering algorithm adjustments for the columns widget
The columns widget's children are supplied along with a dimension
expressing how they are to be rendered, relative to each other. It's
possible to supply a dimension that is both a relative weight and a
maximum number of units at the same time. The idea is that the child
should be rendered with a number of columns determined by the weight
relative to others and the overall number of columns, but that the value
used need not exceed the maximum. If it does, those extra columns are
added back into the pool to be shared among the other children with
relative layouts. The idea is to provide more space for columns that
need it, if particular columns can express an upper limit on what they
need to fully render their content.

The algorithm did not work sensibly when all relatively rendered columns
also had maximums expressed, and the space available exceeded all
maximums. Under those conditions, any extra columns should just be
shared out among all relatively rendered children, even though they
don't need it. These changes and tests capture that behavior.