Fix a slicing bug in the specialized terminal canvas
The ICanvas interface provides a Duplicate() function to make a copy of
the receiver. The terminal uses its own canvas implementation, which
itself uses a viewport over gowid.Canvas, to provide for
scrolling. These types are embedded, and so if the user invokes
Duplicate() on a term canvas, its implementation is provided by
gowid.Canvas.Duplicate(), and the type returned is *gowid.Canvas. This
is a pretty silly error on my part. The fix is to provide dedicated
Duplicate() implementations for both the terminal canvas and the
viewport canvas.