Fixes a styling bug in the dialog widget
Gowid distinguishes between a cell style that is set to off, and a cell
style that expresses no setting at all. That allows for styles to be
combined when layered, so that a lower cell's preference can be used
when an upper cell expresses no style preference. The dialog widget sets
some default styles for use when the user provides none; but if the
underlying widget styles some cells beneath where the dialog would be
rendered, these styles bleed through - by design - to the rendered
canvas. I noticed this in the context of rendering a dialog on top of a
terminal. The intention of the dialog is to cover completely what's
underneath, so I have adjusted the default styles to explicitly set "no
style" for the top layer.