Additions to the standard gowid dialog widget
This changes adds the following:

(1) real "modal" support!  If the Modal option is supplied, then gowid
will not pass user input control out of the dialog widget back to the
parent indicating that the widget did not handle the input - that is,
the widget will always handle the input, whether it processed the
key/mouse event or not.

(2) the ability, via the Tab key, to change focus from the standard
button row to the main dialog content and back again.