Refine the calculation of a subsize used in the vpadding widget
If an app creates a widget like this:

vpadding.New(w, VAlignTop{}, RenderWithUnits{U: 10})

and that widget is rendered in a box context e.g.

RenderBox{C: 5, R: 6}

the w widget will now be rendered with a size of

RenderBox{C: 5, R: 6} and not RenderBox{C: 5, R: 10}

This is a more useful approach, especially when used with widgets like
overlays that would normally want to constrain their subwidgets to
render with dimensions that fit within the overall viewport, rather than
exceeding its bounds (and then being trimmed to fit).