A simple text widget with splitting on whitespace
This is intended for simple use, at the moment, and does not support
styling or any other useful features. Supply it with a string, and the
widget will break the string at whitespace, and lay out the widget in a
flow context according to the specified number of columns, and taking as
much vertical space as it needs. You could use this for displaying
chunks of text that benefit from whitespace-aware splitting when the
output region changes. See the test cases for some example behavior.