Bug fix - when switching screens, deactivate the old screen
I found that when not doing this, the old screen would continue to
consume keypresses for a period of time, presumably competing with the
new screen for reads on the tty until the old one is garbage
collected. I may need to consider how ownership of screens works,
especially if the gowid application is passing in a screen that it
considers itself the owner of - can gowid call Fini() on it,
legitimately?