Respect GOWID_TTY variable
I introduced this variable as a workaround for allowing termshark, a
gowid application, to request to run on a terminal other than
/dev/tty. I didn't change any tcell signatures, but modified a tcell
function to check for a tty in this variable, and then had termshark use
a replace directive to use my fork of tcell. This approach is far from
ideal because (a) using an env var rather than a function signature is
pretty opaque and (b) replace directives + go install, a command I would
like to be able to use as a one-line instruction for installing
termshark. But now that @hinshun has ported gowid to use tcell/v2, I can
eliminate this one use of my gcla/tcell fork because tcell/v2 now
supports opening a screen on a specified tty. For now, I'll just make
that available via the GOWID_TTY variable; in the future, I can expose
this properly via a gowid API.