Support bracketed paste in gowid apps and in the terminal widget
The tcell library supports passing "bracketed paste" events to its
applications. If the tcell app is running in a terminal that supports
it, and the user pastes into their terminal app, tcell will present it
to the application as a tcell.EventPaste event. Gowid will now pass that
event through to widgets so they can make use of it if they need to. By
default the feature is off, but you can enable it by adding the
EnableBracketedPaste field to the AppArgs struct when building an App.

The Edit widget has been modified now to save up regular visible input
when it receives an EventPaste start; then all saved input is emitted at
once, and the widget re-rendered, when the end event is received.

The terminal widget itself now supports an EnableBracketedPaste field in
its widget constructor. If enabled, and if the widget receives a
tcell.EventPaste event, it will translate that to a byte-sequence that
corresponds to the widget's terminfo PasteStart attribute, and send it
to the widget's pseudo-tty. If terminfo does not have such an attribute,
the widget follow's tcell's lead and uses a standard xterm sequence.