Extensions to the terminal widget
Issue #44 requests scrollbar support in the terminal
widget (https://github.com/gcla/gowid/issues/44). This change adds
scrollbar support via a parameter at widget construction time. It also
adds an API to enable or disable the scrollbar at runtime. Added in this
commit, I've provided simple support for adding processing when the
meta-key for the terminal is pressed. This means the widget user can
extend the behaviour of the terminal's meta-key.