Frame widget bug fix for missing constructor attribute
If the New() form of the frame widget constructor is used, and the Frame
attribute is not populated in the Options provided, the widget is
incorrectly rendered. This also means that events inside the frame are
incorrectly offset vertically by 1. This change ensures a legitimate
frame is used, if one is not provided.