Experimental double-click support for the button widget
For example:

  w := button.New(tw, button.Options{
          Decoration:       button.NormalDecoration,
          DoubleClickDelay: 1 * time.Second,
  })

If DoubleClickDelay is omitted, it is set to a default of 0.5s. Handlers
can be added with the new

  OnDoubleClick(..)

API present on the button widget. If the user interacts with the button
in a way that triggers the double-click callback, and if at least one
handler is registered, then the Click(..) handler is suppressed.