Reconsider interpretation of "Color00" when making a color
Gowid provides some convenience functions for making colors from
strings. One such function turns a string name into a TCellColor value,
which can then be used in any Gowid IColor context. These TCellColor
values are fairly direct mappings to the types that tcell itself
provides for colors. As well as translating the descriptive names - like
"burlywood" and "firebrick" - from tcell's color.go, I also match
strings "ColorXY" where XY is a 2-byte hex value and translate them
directly to the tcell Color type based on the fact the underlying value
is a 64-bit int. In the Gowid move from tcell v1 to v2, this function
set an interpretation of Color00 to mean the terminal's default color,
but mapping to tcell's color.go, it would be more consistent for Color00
to mean black. This assumption was used in termshark's default 8-color
theme - that Color00 was black. Without this fix, termshark will
represent its blacks as "terminal default" when TERM=xterm, and the
content may not be readable.