A depth-first search algorithm for the Gowid tree type
Pass in anything implementing tree.IModel and a predicate that expects a
tree node, and the function will return a tree.IPos. The interface will
be nil if nothing matches, otherwise it will be the first matching
position in the tree.

I'll be using this in termshark's upcoming packet-search feature for
searching within the packet "structure". The structure is stored as a
tree within termshark.