An extension to the pile widget's behavior with zero-height widgets
Most gowid widgets render to a canvas with a non-zero height, but
sometimes it's the case that a widget has zero rows - for example the
menu.SiteWidget. The canvas it generates takes no space but does carry
coordinates which are translated during the rendering process and used
by an overlay to place the UI's menus in the intended spot on screen. If
a pile is created where the last item is a menu site, without this
change, the menu does not open because the pile widget terminates
without including the last canvas, even though the addition of a
zero-row canvas would not cause the resulting pile's size to exceed its
rendering constraints. This change causes such zero-row canvases to be
included in the rendered pile widget.