Silly error in the edit widget led to losing cursor control
The UserInput function for the edit widget was returning false even in
cases where it was handling the input. If e.g. edit was embedded in a
column widget with selectable widgets either side, a left/right cursor
press inside widget would move the widget's cursor, but also bubble up
to the column widget and cause it to change the focus widget from edit
to one of its neighbors.