Reduce use of legacy GOWID_TTY variable
This variable was useful before tcell had the ability to construct a
screen with a given tty. I used a fork of tcell that used this variable
internally, and didn't modify any of the tcell APIs. Now tcell/v2 allows
a tty to be specified, so now I am adding a parameter to the App
constructor to provide a tty, then using that preferably.