An option to the overlay widget to apply upper style unilaterally
When merging a lower canvas L beneath an upper canvas U, the following
rules apply:

- the returned canvas uses U's rune if set, otherwise L's
- the returned canvas uses U's color(s) if set, otherwise L's
- for each of the various style types S (bold, underline, etc)
  - the return canvas uses U's pref for S, if provided, otherwise L's

The main point here is that the upper canvas U has to affirmatively
disable a particular style for L's setting not to take effect. This
provides for the most flexibility, allowing overlays to combine styles,
but provides a default that most applications probably wouldn't
want. Applications that e.g. display a dialog widget would be better
served with a default that uses the upper canvas's style.

This changes adds an option to the overlay widget to unilaterally choose
the upper canvas's style. There are some test cases too.