Tweak the way the terminal widget ends scrolling
Thanks to @Peter2121 for this implementation!

Prior to this change, if the user hit the terminal hot-key and scrolled
back up then back down to the prompt again, a further keypress - q or Q
- would still be required to end scroll-mode. This brings the terminal
widget more in-line with other terminal implementations in that
scroll-mode is now ended automatically if the user scrolls back to the
prompt; no extra keypress needed by default.

I have added a new option to the terminal Options struct to preserve the
legacy behavior, if you need it - KeyPressToEndScrollMode (default
false).