Optimize the terminal widget's rendering behavior
Prior to this change, anything arriving on the terminal widget's
descriptor would be parsed and would result in the terminal being
redrawn. But many terminal instructions received this way don't change
the content of the terminal, and so a re-render isn't necessary. This
change refactors the terminal sequence parsing to indicate whether a
redraw is necessary afterwards.