Allow for manual resizing of the terminal
The default is off, meaning the current behavior is preserved - if the
terminal widget is rendered resulting in a new height or width from the
previous render, the underlying terminal is resized too. But, sometimes
it's useful to control this manually e.g. if you are displaying the same
terminal widget twice (mirrored) in the same application, there will
otherwise be a repeating sequence of rerenders as the terminal is
redrawn with different bounding boxes.

The new IHotKeyInputFnExt interface allows for the current rendering
dize to be provided to the hotkey input functions too, should this
information be needed (e.g. to manually resize the terminal)